Moved object type fetching from home view to application start
